requirements.txt
langchain
langchain-openai
langchain-community
langgraph
wikipedia
typing_extensions
python-dotenv